+++ board/pedal/main.c
@@ -8,6 +8,7 @@
 
 #include "drivers/registers.h"
 #include "drivers/interrupts.h"
+#include "drivers/timers.h"
 #include "drivers/llcan.h"
 #include "drivers/llgpio.h"
 #include "drivers/adc.h"
