+++ board/stm32fx/stm32fx_config.h
@@ -27,6 +27,9 @@
 
 #define MICROSECOND_TIMER TIM2
 
+#define INTERRUPT_TIMER_IRQ TIM6_DAC_IRQn
+#define INTERRUPT_TIMER TIM6
+
 #define PROVISION_CHUNK_ADDRESS 0x1FFF79E0
 #define SERIAL_NUMBER_ADDRESS 0x1FFF79C0
